🎨 meilleure séparation des responsabilités entre controller et service
