+++ services/category/package.json
@@ -14,7 +14,8 @@
     "@prisma/client": "^5.21.1",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
-    "express": "^4.21.1"
+    "express": "^4.21.1",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@babel/core": "^7.22.3",
