+++ services/category/src/controllers/category.controller.ts
@@ -1,26 +1,79 @@
 import { CategoryService } from '../services/category.service';
 import { Request, Response } from 'express';
+import { z } from 'zod';
 
 export class CategoryController {
   constructor(private categoryService: CategoryService) {}
 
-  getCategories(req: Request, res: Response) {
-    return this.categoryService.getCategories(req, res);
+  async getCategories(req: Request, res: Response) {
+    try {
+      const categories = await this.categoryService.getCategories();
+      res.status(200).json(categories);
+    } catch (error) {
+      console.error('Error fetching categories:', error);
+      res.status(500).json({ message: 'Error fetching categories' });
+    }
   }
 
-  getCategoryById(req: Request, res: Response) {
-    return this.categoryService.getCategoryById(req, res);
+  async getCategoryById(req: Request, res: Response) {
+    const { id } = req.params;
+    try {
+      const category = await this.categoryService.getCategoryById(id);
+      if (!category) {
+        return res.status(404).json({ message: 'Category not found' });
+      }
+      res.status(200).json(category);
+    } catch (error) {
+      console.error('Error fetching category:', error);
+      res.status(500).json({ message: 'Error fetching category' });
+    }
   }
 
-  createCategory(req: Request, res: Response) {
-    return this.categoryService.createCategory(req, res);
+  async createCategory(req: Request, res: Response) {
+    const categorySchema = z.object({
+      name: z.string().min(1).max(50),
+    });
+
+    try {
+      const { name } = await categorySchema.parseAsync(req.body);
+      const category = await this.categoryService.createCategory(name);
+      res.status(201).json(category);
+    } catch (error) {
+      if (error instanceof z.ZodError) {
+        return res.status(400).json({ message: error.issues });
+      }
+      console.error('Error creating category:', error);
+      res.status(500).json({ message: 'Error creating category' });
+    }
   }
 
-  updateCategory(req: Request, res: Response) {
-    return this.categoryService.updateCategory(req, res);
+  async updateCategory(req: Request, res: Response) {
+    const { id } = req.params;
+    const categorySchema = z.object({
+      name: z.string().min(1).max(50),
+    });
+
+    try {
+      const { name } = await categorySchema.parseAsync(req.body);
+      const category = await this.categoryService.updateCategory(id, name);
+      res.status(200).json(category);
+    } catch (error) {
+      if (error instanceof z.ZodError) {
+        return res.status(400).json({ message: error.issues });
+      }
+      console.error('Error updating category:', error);
+      res.status(500).json({ message: 'Error updating category' });
+    }
   }
 
-  deleteCategory(req: Request, res: Response) {
-    return this.categoryService.deleteCategory(req, res);
+  async deleteCategory(req: Request, res: Response) {
+    const { id } = req.params;
+    try {
+      await this.categoryService.deleteCategory(id);
+      res.status(204).send();
+    } catch (error) {
+      console.error('Error deleting category:', error);
+      res.status(500).json({ message: 'Error deleting category' });
+    }
   }
 }
