+++ services/category/src/services/category.service.ts
@@ -1,88 +1,47 @@
 import { db } from '../database';
-import { Request, Response } from 'express';
 import { slugify } from '../utils/constants';
 
 export class CategoryService {
-  async getCategories(req: Request, res: Response) {
-    try {
-      const categories = await db.category.findMany();
-      res.status(200).json(categories);
-    } catch (error) {
-      console.error('Error fetching categories:', error);
-      res.status(500).json({ message: 'Error fetching categories' });
-    }
+  async getCategories() {
+    return await db.category.findMany();
   }
 
-  async getCategoryById(req: Request, res: Response) {
-    const { id } = req.params;
-    try {
-      const category = await db.category.findUnique({
-        where: {
-          id,
-        },
-      });
-
-      if (!category) {
-        return res.status(404).json({ message: 'Category not found' });
-      }
-
-      return res.status(200).json(category);
-    } catch (error) {
-      console.error('Error fetching category:', error);
-      res.status(500).json({ message: 'Error fetching category' });
-    }
+  async getCategoryById(id: string) {
+    return await db.category.findUnique({
+      where: {
+        id,
+      },
+    });
   }
 
-  async createCategory(req: Request, res: Response) {
-    const { name } = req.body;
+  async createCategory(name: string) {
     const slug = slugify(name);
-    try {
-      const category = await db.category.create({
-        data: {
-          name,
-          slug,
-        },
-      });
-      res.status(201).json(category);
-    } catch (error) {
-      console.error('Error creating category:', error);
-      res.status(500).json({ message: 'Error creating category' });
-    }
+    return await db.category.create({
+      data: {
+        name,
+        slug,
+      },
+    });
   }
 
-  async updateCategory(req: Request, res: Response) {
-    const { id } = req.params;
-    const { name } = req.body;
+  async updateCategory(id: string, name: string) {
     const slug = slugify(name);
-    try {
-      const category = await db.category.update({
-        where: {
-          id,
-        },
-        data: {
-          name,
-          slug,
-        },
-      });
-      res.status(200).json(category);
-    } catch (error) {
-      console.error('Error updating category:', error);
-      res.status(500).json({ message: 'Error updating category' });
-    }
+    return await db.category.update({
+      where: {
+        id,
+      },
+      data: {
+        name,
+        slug,
+      },
+    });
   }
 
-  async deleteCategory(req: Request, res: Response) {
-    const { id } = req.params;
-    try {
-      await db.category.delete({
-        where: {
-          id,
-        },
-      });
-      res.status(204).send();
-    } catch (error) {
-      console.error('Error deleting category:', error);
-      res.status(500).json({ message: 'Error deleting category' });
-    }
+  async deleteCategory(id: string) {
+    await db.category.delete({
+      where: {
+        id,
+      },
+    });
   }
 }
